There is an error while i was trying to rewrite the whole page
