+++ client/src/app/admin/components/categoryComponents/AddCategory.tsx
@@ -1,0 +1,3 @@
+export default function AddCategory() {
+  return <div></div>;
+}
