+++ client/src/app/admin1/components/Logo.tsx
@@ -1,0 +1,11 @@
+export default function Logo() {
+  return (
+    <div className="flex gap-[8px]">
+      <img src="./images/icons/Logo.png"></img>
+      <div>
+        <h2 className="text-[18px] font-[600]">NomNom</h2>
+        <p className="text-[12px] font-[400] text-[#71717A]">Swift delivery</p>
+      </div>
+    </div>
+  );
+}
