+++ client/src/app/admin1/components/Orders.tsx
@@ -1,0 +1,8 @@
+import { OrdersTable } from "./ordersComponents/OrdersTable";
+export const Orders = () => {
+  return (
+    <div className="bg-[#F4F4F5] w-[100%] h-[100vh]">
+      <OrdersTable />
+    </div>
+  );
+};
