+++ client/src/app/admin1/components/categoryComponents/CategoryFoodCard.tsx
@@ -1,0 +1,3 @@
+export default function CategoryCard() {
+  return <div></div>;
+}
