+++ client/src/app/admin1/components/ordersComponents/Order.tsx
@@ -1,0 +1,3 @@
+export default function Order() {
+  return <div></div>;
+}
