+++ client/src/app/admin1/components/ordersComponents/Orders.tsx
@@ -1,0 +1,3 @@
+export default function Orders() {
+  return <div></div>;
+}
