+++ client/src/app/admin1/components/ordersComponents/OrdersTable.tsx
@@ -1,0 +1,10 @@
+import OrdersTableHeader from "./OrdersTableHeader";
+import Orders from "./Orders";
+export const OrdersTable = () => {
+  return (
+    <div>
+      <OrdersTableHeader />
+      <Orders />
+    </div>
+  );
+};
