+++ client/src/app/admin1/components/ordersComponents/OrdersTableHeader.tsx
@@ -1,0 +1,3 @@
+export default function OrdersTableHeader() {
+  return <div></div>;
+}
