+++ client/src/app/admin1/page.tsx
@@ -1,0 +1,4 @@
+import Container from "./components/Container";
+export default function page() {
+  return <Container />;
+}
