+++ client/src/app/admin1/types.ts
@@ -1,0 +1,12 @@
+export type foodType = {
+  _id: string;
+  foodName: string;
+  price: number;
+  image: string;
+  ingredients: string;
+  category: string;
+};
+export type categoryType = {
+  categoryName?: string;
+  _id?: string;
+};
