+++ client/src/app/globals.css
@@ -18,8 +18,14 @@
     --foreground: #ededed;
   }
 }
-
+html {
+  height: 100%;
+}
 body {
+  min-height: 100%;
+  margin: 0px;
+  padding: 0px;
+  box-sizing: border-box;
   background: var(--background);
   color: var(--foreground);
   font-family: Arial, Helvetica, sans-serif;
